Implemented rebooting via update.jar.
Also, System.out is always redirected to ConsoleOutputTracker regardless
of manual_reconnecting.

+++ src/main/java/me/kagerou/kyoukobot/Update.java
@@ -2,13 +2,15 @@ package me.kagerou.kyoukobot;
 
 import java.io.File;
 import java.lang.ProcessBuilder.Redirect;
+import java.util.ArrayList;
+import java.util.List;
 
 public class Update {
 
 	public static void main(String[] args) {
 		if ((args.length != 2) && (args.length != 3))
 		{
-			System.out.println("Usage: java -jar update.jar tmpfile jarfile [beta]");
+			System.out.println("Usage: java -jar update.jar (tmpfile/reboot) jarfile [beta]");
 			return;
 		}
 		try {
@@ -19,32 +21,40 @@ public class Update {
 			System.out.println("Interrupted!");
 			return;
 		}
-		File tmpFile = new File(args[0]);
 		File jarFile = new File(args[1]);
-		if (!tmpFile.exists())
+		boolean beta = (args.length > 2) && (args[2].equals("beta"));
+		List<String> commands = new ArrayList<String>();
+		commands.add("java");
+		commands.add("-jar");
+		commands.add(args[1]);
+		if (beta)
+			commands.add("beta");
+		if (args[0].equalsIgnoreCase("reboot"))
+			System.out.println("Restarting " + args[1] + "...");
+		else
 		{
-			System.out.println("File " + args[0] + " not found.");
-			return;
-		}
-		if (jarFile.delete())
-			System.out.println("Deleted " + args[1] + " successfully!");
-		else
-			System.out.println("File " + args[1] + " not found.");
-		if (tmpFile.renameTo(jarFile))
-			System.out.println("Renamed " + args[0] + " to " + args[1] + ".");
-		else
-			System.out.println("Failed to rename " + args[0] + " to " + args[1] + ".");
-		try {
-			boolean beta = (args.length > 2) && (args[2].equals("beta"));
-			if (!beta)
-				new ProcessBuilder("java", "-jar", args[1], "updated").redirectOutput(Redirect.INHERIT).redirectError(Redirect.INHERIT).start();
+			File tmpFile = new File(args[0]);
+			if (!tmpFile.exists())
+			{
+				System.out.println("File " + args[0] + " not found.");
+				return;
+			}
+			if (jarFile.delete())
+				System.out.println("Deleted " + args[1] + " successfully!");
 			else
-				new ProcessBuilder("java", "-jar", args[1], "updated", "beta").redirectOutput(Redirect.INHERIT).redirectError(Redirect.INHERIT).start();
+				System.out.println("File " + args[1] + " not found.");
+			if (tmpFile.renameTo(jarFile))
+				System.out.println("Renamed " + args[0] + " to " + args[1] + ".");
+			else
+				System.out.println("Failed to rename " + args[0] + " to " + args[1] + ".");
+			commands.add("updated");
+		}
+		try {
+			new ProcessBuilder(commands).redirectOutput(Redirect.INHERIT).redirectError(Redirect.INHERIT).start();
 		}
 		catch (Exception e)
 		{
 			e.printStackTrace();
 		}
 	}
-
 }
